Added Features, Changes and Fixies
New Features:
In the project, the MVC (Model-View-Controller) pattern has been implemented solely for the PvP game controller, and modifications and alterations have also been introduced.

Added:
A main menu exit button when the game ends. Thus, after the game is finished, the player has two options: play again or exit to the main menu.

Changes:
In the Score section when the game starts, it displays the players' names and their scores. Additionally, if it's a tie, the score doesn't increase.

Visualization of whose turn it is to play. Now, instead of being highlighted in green, it indicates which player's turn it is alongside the score.

Fixed:
The ability to simultaneously change user names has been rectified.